src/PathPlanPrinter/requirements.txt
Pillow==7.1.1
numpy==1.18.2

